RUN: Introduce RsAsyncRunner
