+++ clion/src/191/main/kotlin/org/rust/clion/profiler/RsProfilerRunner.kt
@@ -1,0 +1,19 @@
+/*
+ * Use of this source code is governed by the MIT license that can be
+ * found in the LICENSE file.
+ */
+
+package org.rust.clion.profiler
+
+import com.intellij.profiler.clion.ProfilerExecutor
+import org.rust.cargo.runconfig.RsAsyncRunner
+
+const val ERROR_MESSAGE_TITLE: String = "Profiling is not possible"
+
+class RsProfilerRunner : RsAsyncRunner(ProfilerExecutor.EXECUTOR_ID, ERROR_MESSAGE_TITLE) {
+    override fun getRunnerId(): String = RUNNER_ID
+
+    companion object {
+        const val RUNNER_ID: String = "RsProfilerRunner"
+    }
+}
